+++ Mods ChoGGi/All Sponsors/metadata.lua
@@ -1,0 +1,14 @@
+return PlaceObj("ModDef", {
+	"title", "All Sponsors v0.1",
+	"version", 1,
+	"saved", 1545307200,
+	"image", "Preview.png",
+	"id", "ChoGGi_AllSponsors",
+	"steam_id", "1596612055",
+	"author", "ChoGGi",
+	"lua_revision", LuaRevision,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[If you have three sponsors; this will add all the other sponsors.]],
+})